# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistem Informasi Kepegawaian dan Status Kehadiran API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: getEmployees
      tags: [employees]
      summary: Get all employees
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployee"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
...
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployee"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /documents:
    get:
      operationId: getDocuments
      tags: [documents]
      summary: Get all documents
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [SP3S, SIJ, CUTI, DINAS, SKMJ, SURAT_TUGAS]
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create new document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocument"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    StatusPekerja:
      type: string
      enum: [Organik, TAD, TKJP, Security, Mitra Kerja, Driver, CS, Gardener]
    StatusBekerja:
      type: string
      enum: [Aktif, Mutasi, Pensiun, PHK, Mengundurkan Diri]
    Employee:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        nopek:
          type: string
        foto:
          type: string
          nullable: true
        jabatan:
          type: string
          nullable: true
        departemen:
          type: string
          nullable: true
        fungsi:
          type: string
          nullable: true
        lokasi_kerja:
          type: string
          nullable: true
        nomor_telepon:
          type: string
          nullable: true
        status_pekerja:
          $ref: "#/components/schemas/StatusPekerja"
          nullable: true
        status_bekerja:
          $ref: "#/components/schemas/StatusBekerja"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - nama
        - nopek
        - createdAt
    CreateEmployee:
      type: object
      properties:
        nama:
          type: string
        nopek:
          type: string
        foto:
          type: string
          nullable: true
        jabatan:
          type: string
          nullable: true
        departemen:
          type: string
          nullable: true
        fungsi:
          type: string
          nullable: true
        lokasi_kerja:
          type: string
          nullable: true
        nomor_telepon:
          type: string
          nullable: true
        status_pekerja:
          $ref: "#/components/schemas/StatusPekerja"
          nullable: true
        status_bekerja:
          $ref: "#/components/schemas/StatusBekerja"
          nullable: true
      required:
        - nama
        - nopek
...
    Document:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employee:
          $ref: "#/components/schemas/Employee"
        type:
          type: string
          enum: [SP3S, SIJ, CUTI, DINAS, SKMJ, SURAT_TUGAS]
        nomorSurat:
          type: string
          nullable: true
        perihal:
          type: string
          nullable: true
        tanggal:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        keterangan:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - employeeId
        - type
        - tanggal
        - status
        - createdAt
    CreateDocument:
      type: object
      properties:
        employeeId:
          type: integer
        type:
          type: string
          enum: [SP3S, SIJ, CUTI, DINAS, SKMJ, SURAT_TUGAS]
        nomorSurat:
          type: string
          nullable: true
        perihal:
          type: string
          nullable: true
        tanggal:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        keterangan:
          type: string
          nullable: true
      required:
        - employeeId
        - type
        - tanggal
        - status
...